Got the TensorFlow classifier working. JSON file export for transforms.
Managed to find the error in the TesnorFlow classifier. The module-info
file generates "api" as the name for the djl library. This has now been
fixed and will be updated in the next djl release. for now not using
module-info file. Also added the ability to export transform settings to
a JSON file.