Moved from JFrog to GitHub
jpam started off on BinTray which was shut down by the parent company JFrog. It then moved to JFrog's alternative which was over complicated and shut down accounts of they were not accessed for a few weeks. I have therefore migrated the maven repo to GitHub - the JFrog repo will be deleted in a few weeks.